% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{areColors}
\alias{areColors}
\title{Quick utility to check if given character vector is colors
Thanks to Stackoverflow: http://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation}
\usage{
areColors(x)
}
\arguments{
\item{x}{character vector to check}
}
\value{
boolean whether character vector is colors
}
\description{
Quick utility to check if given character vector is colors
Thanks to Stackoverflow: http://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation
}
\keyword{internal}
