% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{armaCor}
\alias{armaCor}
\title{armaCor - matrix column correlations}
\usage{
armaCor(mat)
}
\arguments{
\item{mat}{matrix, whose columns will be}
}
\description{
similar to cor() call, will calculate correlation between matrix columns
}
