% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{basicP2proc}
\alias{basicP2proc}
\title{Perform basic pagoda2 processing}
\usage{
basicP2proc(
  cd,
  n.cores = 1,
  batch = NULL,
  n.odgenes = 3000,
  nPcs = 100,
  k = 30,
  perplexity = 50,
  log.scale = TRUE,
  trim = 10,
  keep.genes = NULL,
  min.cells.per.gene = 0,
  min.transcripts.per.cell = 100,
  get.largevis = TRUE,
  get.tsne = TRUE,
  make.geneknn = TRUE
)
}
\arguments{
\item{cd}{count matrix, rows are genes, columns are cells}

\item{n.cores}{number of cores to use}

\item{batch}{optional batch factor (default=NULL)}

\item{n.odgenes}{number of top overdispersed genes to use (dfault=3e3)}

\item{nPcs}{number of PCs to use}

\item{k}{default number of neighbors to use in kNN graph}

\item{perplexity}{perplexity to use in generating tSNE and largeVis embeddings (default=50)}

\item{log.scale}{whether to use log scale normalization (default=T)}

\item{trim}{number of cells to trim in winsorization (default=10)}

\item{keep.genes}{optional set of genes to keep from being filtered out (even at low counts, default=NULL)}

\item{min.cells.per.gene}{minimal number of cells required for gene to be kept (unless listed in keep.genes)}

\item{min.transcripts.per.cell}{minimumal number of molecules/reads for a cell to be admitted}

\item{get.largevis}{whether to caluclate largeVis embedding}

\item{get.tsne}{whether to calculate tSNE embedding}

\item{make.geneknn}{whether pre-calculate gene kNN (for gene search)}
}
\value{
a new pagoda2 object
}
\description{
adjust variance, calculate pca reduction
make knn graph, identify clusters with infomap, multilevel and walktrap and make
largeVis and tSNE embeddings
}
