% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{basicP2web}
\alias{basicP2web}
\title{Generate a p2 web application from a pagoda2 object}
\usage{
basicP2web(p2, app.title = "Pagoda2", extraWebMetadata = NULL,
  n.cores = 4)
}
\arguments{
\item{p2}{pagoda2 application object}

\item{app.title}{name of application as displayed in the browser title}

\item{extraWebMetadata}{additional metadata generated by p2.metadata.from.fractor}

\item{n.cores}{number of cores to use for differential expression calculation}
}
\value{
a pagoda2 web object
}
\description{
Generate a p2 web application from a pagoda2 object
}
