% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{calcMulticlassified}
\alias{calcMulticlassified}
\title{returns a list vector with the number of multiclassified cells}
\usage{
calcMulticlassified(sel)
}
\arguments{
\item{sel}{a pagoda2 selection as genereated by readPagoda2SelectionFile}
}
\description{
returns a list vector with the number of cells that are
present in more than one selections in the provided p2 selection object
}
