% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{cellsPerSelectionGroup}
\alias{cellsPerSelectionGroup}
\title{get the number of cells per selection group}
\usage{
cellsPerSelectionGroup(selection)
}
\arguments{
\item{selection}{a pagoda2 selection list}
}
\value{
a named vector of cell numbers in each grous
}
\description{
get the number of cells in each selection group
}
