% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cldend2array}
\alias{cldend2array}
\title{Translate cell cluster dendrogram to an array, one row per node with 1/0 cluster membership}
\usage{
cldend2array(d, cells = NULL)
}
\arguments{
\item{d}{cell cluster dendrogram}

\item{cells}{character vector of cells (default=NULL). If NULL, determine the total order of cells with unlist(dendrapply(d, attr, 'cells'))}
}
\value{
array with one row per node with 1/0 cluster membership
}
\description{
Translate cell cluster dendrogram to an array, one row per node with 1/0 cluster membership
}
\keyword{internal}
