% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{cleanupGraph}
\alias{cleanupGraph}
\title{Clean up an inter-sample graph}
\usage{
cleanupGraph(g, min.neigh.con = 1, show.progress = TRUE)
}
\arguments{
\item{g}{the graph to cleanup}

\item{min.neigh.con}{the number of neighbour matches to require}
}
\value{
a new cleaned up graph
}
\description{
Clean up an inter-sample graph
}
