% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-pagoda1.R
\name{collapse.aspect.clusters}
\alias{collapse.aspect.clusters}
\title{Collapse aspect patterns into clusters}
\usage{
collapse.aspect.clusters(d, dw, ct, scale = TRUE, pick.top = FALSE)
}
\arguments{
\item{d}{matrix of normalized aspect patterns (rows: significant aspects, columns: cells), normally the output $xv in 'tamr', the combined pathways that show similar expression patterns}

\item{dw}{corresponding weight matrix to parameter 'd'}

\item{ct}{clusters, the output of fastcluster::hclust()}

\item{scale}{boolean Whether to scale aspects (default=TRUE)}

\item{pick.top}{boolean Whether to pick top aspects (default=FALSE)}
}
\value{
list of clusters from matrix of normalized aspect patterns and clusters from the corresponding weight matrix
}
\description{
Collapse aspect patterns into clusters
}
\keyword{internal}
