% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{compareClusterings}
\alias{compareClusterings}
\title{Compare two different clusterings provided as factors by plotting a normalised heatmap}
\usage{
compareClusterings(cl1, cl2, filename = NA)
}
\arguments{
\item{cl1}{clustering 1, a named factor}

\item{cl2}{clustering 2, a named factor}

\item{filename}{an optional filename to save the plot instead of displaying it, will be passed to pheatmap (default=NA)}
}
\value{
invisible summary table that gets plotted
}
\description{
Compare two different clusterings provided as factors by plotting a normalised heatmap
}
