% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{diffExprOnP2FromWebSelection}
\alias{diffExprOnP2FromWebSelection}
\title{Perform differential expression on a p2 object given a
set of web selections and two groups to compare}
\usage{
diffExprOnP2FromWebSelection(p2, sel, group1name, group2name)
}
\arguments{
\item{p2}{a pagoda2 object}

\item{sel}{a web selection read with readPagoda2SelectionFile()}

\item{group1name}{the name of the first selection}

\item{group2name}{the names of the second selection}
}
\description{
Perform differential expression on a p2 object given a
set of web selections and two groups to compare
}
\keyword{internal}
