% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{diffExprOnP2FromWebSelectionOneGroup}
\alias{diffExprOnP2FromWebSelectionOneGroup}
\title{Perform differential expression on a p2 object given a
set of web selections and one group to compare against everything else}
\usage{
diffExprOnP2FromWebSelectionOneGroup(p2, sel, groupname)
}
\arguments{
\item{p2}{a pagoda2 object}

\item{sel}{a web selection read with readPagoda2SelectionFile()}

\item{groupname}{name of group to compare to everything else}
}
\description{
Perform differential expression on a p2 object given a
set of web selections and one group to compare against everything else
}
\keyword{internal}
