% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{extendedP2proc}
\alias{extendedP2proc}
\title{Perform extended pagoda2 processing}
\usage{
extendedP2proc(p2, n.cores = 20, organism = "hs")
}
\description{
generate organism specific GO environment and
calculate pathway overdispersion
}
