% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{factorFromP2Selection}
\alias{factorFromP2Selection}
\title{returns a factor of cell membership from a p2 selection}
\usage{
factorFromP2Selection(sel, use.internal.name = FALSE, flatten = FALSE)
}
\arguments{
\item{flatten}{ignores multiclassified cells, overwriting randomly}
}
\description{
returns a factor of cell membership from a p2 selection object
the factor only includes cells present in the selection. If the selection
contains multiclassified cells an error is raised
}
