% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{factorToP2selection}
\alias{factorToP2selection}
\title{converts a factor to a p2 selection object}
\usage{
factorToP2selection(cl, col = NULL)
}
\arguments{
\item{col}{names vector of colors}
}
\value{
a p2 selection object (list)
}
\description{
converts a names factor to a p2 selection object
if colors are provided it assigns those, otherwise uses a rainbow palette
}
