% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{generateClassificationAnnotation}
\alias{generateClassificationAnnotation}
\title{Given a cell clustering (partitioning) and a set of user provided selections
generate a cleaned up annotation of cluster groups that can be used for classification}
\usage{
generateClassificationAnnotation(clustering, selections)
}
\arguments{
\item{clustering}{a factor that provides the clustering}

\item{selections}{a p2 selection object that provided by the web interfact user}
}
\value{
a named factor that can be used for classification
}
\description{
Given a cell clustering (partitioning) and a set of user provided selections
generate a cleaned up annotation of cluster groups that can be used for classification
}
