% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-scoring.R
\name{get.control.geneset}
\alias{get.control.geneset}
\title{Get a control geneset for cell scoring using the method described in
Puram, Bernstein (Cell, 2018)}
\usage{
get.control.geneset(data, signature, n.bins = 25, n.genes.per.bin = 100)
}
\arguments{
\item{data}{matrix of expression, rows are cell, columns are genes}

\item{signature}{character vector The signature to evaluate, a character vector of genes}

\item{n.bins}{numeric Number of bins to put the genes in (default=25)}

\item{n.genes.per.bin}{numeric Number of genes to get from each bin (default=100)}
}
\value{
a character vector that can be used as a background signature
}
\description{
Get a control geneset for cell scoring using the method described in
Puram, Bernstein (Cell, 2018)
}
