% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.de.geneset}
\alias{get.de.geneset}
\title{Generate differential expression genesets for the web app given a cell grouping}
\usage{
get.de.geneset(pagObj, groups, prefix = "de_")
}
\arguments{
\item{pagObj}{pagoda object}

\item{groups}{a named factor to do the de by}

\item{prefix}{a character  prefix to assign to genesets generated}
}
\description{
Generate differential expression genesets for the web app given a cell grouping by
Calculating de sets between every cell set and everything else individually
}
