% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{get.joint.clustering}
\alias{get.joint.clustering}
\title{Obtain joint clustering between two samples by constructing MNN network
and calling clusters on that}
\usage{
get.joint.clustering(r.n, k = 30,
  community.detection.method = walktrap.community, min.group.size = 10)
}
\arguments{
\item{r.n}{a named list of pagoda2 apps}

\item{k}{number of KNN to look for}

\item{community.detection.method}{one of the igraph package community detection methods (or compatible fn)}

\item{min.group.size}{minimum group size to keep after matching}
}
\value{
a named (by cell name) factor of groups
}
\description{
Obtain joint clustering between two samples by constructing MNN network
and calling clusters on that
}
