% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{getIntExtNamesP2Selection}
\alias{getIntExtNamesP2Selection}
\title{Get a mapping form internal to external names for the specified selection object}
\usage{
getIntExtNamesP2Selection(x)
}
\arguments{
\item{x}{p2 selection object}
}
\description{
Get a mapping form internal to external names for the specified selection object
}
