% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{getJointClustering}
\alias{getJointClustering}
\title{Obtain joint clustering between multiple pagoda2 applications}
\usage{
getJointClustering(r.n, k = 30,
  community.detection.method = walktrap.community, min.group.size = 10,
  ncomps = 100)
}
\arguments{
\item{r.n}{a named list of pagoda2 apps}

\item{k}{number of KNN to look for}

\item{community.detection.method}{one of the igraph package community detection methods (or compatible fn)}

\item{min.group.size}{minimum group size to keep after matching}

\item{ncomps}{number of components to use}
}
\value{
a named (by cell name) factor of groups
}
\description{
obtain joint clustering between multiple pagoda2 applications
by identifying mutual nearest neighbours pairwise and constructing
a network on which traditional cluster identification algorithms are used to identify clusters
}
