% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{getJointClustering2}
\alias{getJointClustering2}
\title{Devel version of getJointClustering}
\usage{
getJointClustering2(r.n, k = 30,
  community.detection.method = walktrap.community, min.group.size = 10,
  ncomps = 100, include.sample.internal.edges = TRUE, mnn.edge.weight = 1,
  internal.edge.weight = 1, extra.info = F, mnnPairFunction = NULL)
}
\description{
Devel version of getJointClustering
}
