% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{getJointClusterMarkerGenes}
\alias{getJointClusterMarkerGenes}
\title{Get list of genes for each cluster that are consistently diff expressed
in all the apps in that cluster}
\usage{
getJointClusterMarkerGenes(applist, jc)
}
\arguments{
\item{applist}{list of apps}

\item{jc}{joint clusters}
}
\value{
a list of gene vectors one entry for each cluster
}
\description{
Get list of genes for each cluster that are consistently diff expressed
in all the apps in that cluster
}
