% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{getMNNforP2pair}
\alias{getMNNforP2pair}
\title{Get mutual neighbours for 2 pagoda2 objects}
\usage{
getMNNforP2pair(r1, r2, var.scale = T, k = 30, log.scale = T,
  center = T, verbose = T)
}
\arguments{
\item{r1}{first pagoda object}

\item{r2}{second pagoda object}

\item{var.scale}{option to scale for variance}

\item{k}{number of nearest neighbours}

\item{log.scale}{work on logarithmic scale}

\item{center}{center the data}

\item{verbose}{verbosity level}
}
\value{
a data frame with all the pairs of the mutual neighbours
}
\description{
Get mutual neighbours for 2 pagoda2 objects
}
