% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{getMNNforP2pairCustom}
\alias{getMNNforP2pairCustom}
\title{A replacement function of getMNNforP2pairCustom}
\usage{
getMNNforP2pairCustom(r1, r2, var.scale = T, k = 30, log.scale = T,
  center = T, verbose = T, ncomps = 100, plot.projection = F)
}
\description{
A replacement function of getMNNforP2pairCustom
}
