% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{hierDiffToGenesets}
\alias{hierDiffToGenesets}
\title{Converts the output of hierarchical differential expression aspects
into genesets that can be loaded into a 'pagoda2' web app to retrive the genes
that make the geneset interactively}
\usage{
hierDiffToGenesets(output)
}
\arguments{
\item{output}{output of getHierarchicalDiffExpressionAspects}
}
\value{
a geneset that can be loaded into p2 web genesets
}
\description{
Converts the output of hierarchical differential expression aspects
into genesets that can be loaded into a 'pagoda2' web app to retrive the genes
that make the geneset interactively
}
