% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{identifyCellsGSVDMNN}
\alias{identifyCellsGSVDMNN}
\title{Identify cells in GSVD space with MNN summarisation}
\usage{
identifyCellsGSVDMNN(referenceP2, r2, referenceP2labels, var.scale = T,
  k = 30, log.scale = T, center = T, verbose = T, extra.info = F)
}
\arguments{
\item{referenceP2}{pagoda2 object to use as reference for the annotation}

\item{r2}{pagoda2 object to the annotation of}

\item{var.scale}{perform variance scaling (according to p2 variance estimate)}

\item{k}{number of nearest neighbours to look for (before filtering for mutual)}

\item{log.scale}{transform to log scale}

\item{center}{perform centering}

\item{verbose}{print progress messages}

\item{extra.info}{return extra information from the run, this is currently the gsvd decomposition object}

\item{referenceP2lables}{a factor providing labels for every cell in reference P2}
}
\description{
Identify cells in GSVD space with MNN summarisation
}
