% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{identifyCellsGSVDMNNmulti}
\alias{identifyCellsGSVDMNNmulti}
\title{identify cells using teh identifyCellsGSVDMNN function using multiple input
reference sets}
\usage{
identifyCellsGSVDMNNmulti(referencesets, annotset, clustersOrig)
}
\arguments{
\item{referencesets}{input datasets (p2 objects in a names list)}

\item{annotset}{p2 dataset to be annotated}

\item{clustersOrigin}{a names factor with cluster assignments for referenceset cells}
}
\value{
a named factor with labells for annotset
}
\description{
identify cells using teh identifyCellsGSVDMNN function using multiple input
reference sets
}
