% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{identifyCellsGSVDRF}
\alias{identifyCellsGSVDRF}
\title{Classify cells in a pagoda2 application given an annotated pagoda2 object}
\usage{
identifyCellsGSVDRF(referenceP2, r2, referenceP2labels, var.scale = T,
  log.scale = T, center = T, verbose = T, extra.info = F, n.cores = 1,
  n.trees = 1000)
}
\arguments{
\item{referenceP2}{pagoda2 object to use as reference for the annotation}

\item{r2}{pagoda2 object to the annotation of}

\item{var.scale}{perform variance scaling (according to p2 variance estimate)}

\item{center}{perform centering}

\item{verbose}{print progress messages}

\item{extra.info}{return extra information from the run}

\item{n.cores}{number of cores to use}

\item{n.trees}{number of trees to use in total}

\item{referenceP2lables}{a factor providing labels for every cell in reference P2}
}
\description{
Given a reference pagoda2 application an a set of labels for each
cell, classify cells in a new application. This function works by placing both
datasets onto a common GSVD and using random forests in that sapce
}
