% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{identifyGSVDRFMulti}
\alias{identifyGSVDRFMulti}
\title{Classify cells in a pagoda2 application given multiple pagoda2 objects}
\usage{
identifyGSVDRFMulti(referencesets, annotset, clustersOrig, n.trees = 100,
  n.cores = 1)
}
\arguments{
\item{referencesets}{input datasets (p2 objects in a names list)}

\item{annotset}{p2 dataset to be annotated}

\item{n.trees}{number of trees to use per app in total}

\item{n.cores}{number of cores to use for the random forests}

\item{clustersOrigin}{a names factor with cluster assignments for referenceset cells}
}
\value{
a named factor with labells for annotset
}
\description{
This function is similar to identifyCellsGSVDRF() but it can
take as input a list of annotated pagoda2 objects. It then classifies cells
against all applications and summarises the results.
}
