% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{jw.disR}
\alias{jw.disR}
\title{Jaro-Winkler distance}
\usage{
jw.disR(x, y)
}
\arguments{
\item{x}{numeric vector first input}

\item{y}{numeric vector second input}
}
\description{
Jaro-Winkler distance
}
\keyword{internal}
