% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{make.p2.app}
\alias{make.p2.app}
\title{Generate a Rook Server app from a pagoda2 object}
\usage{
make.p2.app(r, dendrogramCellGroups, additionalMetadata = list(), geneSets,
  show.depth = T, show.batch = T, show.clusters = T,
  appname = "Pagoda2 Application", innerOrder = NULL, orderDend = FALSE)
}
\arguments{
\item{r}{pagoda2 object}

\item{additionalMetadata}{a list of metadata other than depth, batch and cluster that are automatically added}

\item{geneSets}{a list of genesets to show}

\item{show.depth}{logical, include depth as a metadata row}

\item{show.batch}{logical, include batch as a metadata row}

\item{show.clusters}{logical, include clusters as a metadata row}

\item{appname}{application name}

\item{innerOrder}{Ordering of cells inside the clusters provided in dendrogramCellGroups. This should be one of "odPCA", "reductdist", "graphbased", "knn". Defaults to NULL}

\item{dendrogramCelllGoups}{a named factor of cell groups, used to generate the main dendrogram, limits zoom in}
}
\value{
a pagoda2 web object that presents a Rook compatible interface
}
\description{
Generates a pagoda 2 web object from pagoda2 object by automating steps that most
users will want to run. This function is a wrapper about the pagoda2 web constructor. Advanced users
may wish to use use the constructor directly
}
