% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Mode}
\alias{Mode}
\title{Return the mode of a vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{the vector to return the mode of}
}
\value{
the mode elements
}
\description{
Return the mode of a vector
}
\keyword{internal}
