% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{p2.generate.dr.go}
\alias{p2.generate.dr.go}
\title{Generate a GO environment for human for overdispersion analysis for the the back end}
\usage{
p2.generate.dr.go(r)
}
\arguments{
\item{r}{pagoda2 object}
}
\value{
a GO environment object
}
\description{
Generate a GO environment for human for overdispersion analysis for the the back end
}
\examples{
\donttest{
cm <- readRDS(system.file("extdata", "sample_BM1.rds", package="pagoda2"))
p2 <- basicP2proc(cm)
p2.generate.dr.go(p2)
}
}
