% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebHelpers.R
\name{p2.generate.go.web.fromGOEnv}
\alias{p2.generate.go.web.fromGOEnv}
\title{Generates GO annotation for the web object from the GO environment used for enrichment analysis}
\usage{
p2.generate.go.web.fromGOEnv(go.env)
}
\arguments{
\item{go.env}{GO enviroment generated with p2.generate.go}
}
\description{
Generates GO annotation for the web object from the GO environment used for enrichment analysis
}
\keyword{internal}
