% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebHelpers.R
\name{p2.generate.go.web}
\alias{p2.generate.go.web}
\title{Generates human go annotation for the web object for any species}
\usage{
p2.generate.go.web(
  gene.names,
  egALIAS2EG = NULL,
  egGO2ALLEGS = NULL,
  n.cores = 1
)
}
\arguments{
\item{gene.names}{a character vector of genes to include}
}
\description{
generates a humna go annotation for the pagoda2 web object
}
