% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebHelpers.R
\name{p2.generate.human.go.web}
\alias{p2.generate.human.go.web}
\title{Generates human go annotation for the web object}
\usage{
p2.generate.human.go.web(gene.names, n.cores = 1)
}
\arguments{
\item{gene.names}{a character vector of genes to include}

\item{n.cores}{numeric Number of cores to use (default=1)}
}
\description{
generates a humna go annotation for the pagoda2 web object
}
