% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{p2.generate.mouse.go}
\alias{p2.generate.mouse.go}
\title{Generate a GO environment for mouse for overdispersion analysis for the the back end}
\usage{
p2.generate.mouse.go(r)
}
\arguments{
\item{r}{a 'Pagoda2' object}
}
\value{
a GO environment object
}
\description{
Generate a GO environment for mouse for overdispersion analysis for the the back end
}
