% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebHelpers.R
\name{p2.generate.mouse.go.web}
\alias{p2.generate.mouse.go.web}
\title{Generates mouse go annotation for the web object}
\usage{
p2.generate.mouse.go.web(gene.names, n.cores = 1)
}
\arguments{
\item{gene.names}{a character vector of genes to include}
}
\description{
generates a humna go annotation for the pagoda2 web object
}
