% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{p2.toweb.hdea}
\alias{p2.toweb.hdea}
\title{Generate a pagoda2 web object from a pagoda2 object using hierarchical differential expression}
\usage{
p2.toweb.hdea(p2, title = "")
}
\arguments{
\item{p2}{p2 object}

\item{title}{name of the pagoda object}
}
\description{
Generate a pagoda2 web object from a pagoda2 object using hierarchical differential expression
}
