% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p1app4p2.R
\name{p2app4conos}
\alias{p2app4conos}
\title{Utility function to generate a pagoda2 app from a conos object}
\usage{
p2app4conos(
  conos,
  cdl = NULL,
  metadata = NULL,
  filename = "conos_app.bin",
  go.env = NULL,
  organism = "hs",
  save = TRUE,
  n.cores = 2,
  cell.subset = NULL,
  max.cells = Inf,
  additional.embeddings = NULL,
  test.pathway.overdispersion = TRUE,
  return.details = FALSE
)
}
\arguments{
\item{conos}{Conos object}

\item{cdl}{list Optional list of raw matrices (so that gene merging doesn't have to be redone) (default=NULL)}

\item{metadata}{list Optional list of (named) metadata factors (default=NULL)}

\item{filename}{string Name of the *.bin file to seralize for the pagoda2 application if save=TRUE (default='conos_app.bin')}

\item{go.env}{GO environment for the organism of interest (default=NULL)}

\item{organism}{string Organism of interest, either 'hs' (Homo sapiens) or 'mm' (Mus musculus, i.e. mouse) (default='hs')}

\item{save}{boolean Save serialized *bin file specified in filename (default=TRUE)}

\item{n.cores}{integer Number of cores (default=2)}

\item{cell.subset}{string Cells to subset with the conos embedding conos$embedding. If NULL, uses all cells via rownames(conos$embedding) (default-NULL)}

\item{max.cells}{numeric Limit to the cells that are included in the conos. If Inf, there is no limit (default=Inf)}

\item{additional.embeddings}{list Additional embeddings to add to conos for the pagoda2 app (default=NULL)}

\item{test.pathway.overdispersion}{boolean Find all IDs using GO category against either org.Hs.eg.db or org.Mm.eg.db (default=TRUE)}

\item{return.details}{boolean If TRUE, return list of p2 application, pagoda2 object, list of raw matrices, and cell names. If FALSE, simply return pagoda2 app object. (default=FALSE)}
}
\value{
pagoda2 app object
}
\description{
Utility function to generate a pagoda2 app from a conos object
}
