% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p1app4p2.R
\docType{class}
\name{p2ViewPagodaApp-class}
\alias{p2ViewPagodaApp-class}
\alias{p2ViewPagodaApp}
\title{Modified PAGODA1 app (from scde) for browsing pagoda2 results}
\description{
Modified PAGODA1 app (from scde) for browsing pagoda2 results
}
\section{Fields}{

\describe{
\item{\code{results}}{result object returned by \code{scde.expression.difference()}. Note to browse group posterior levels, use \code{return.posteriors = TRUE} in the \code{scde.expression.difference()} call.}

\item{\code{tam}}{Combined pathways that are driven by the same gene sets}

\item{\code{genes}}{List of genes to display in the Detailed clustering panel}

\item{\code{pathways}}{character vector Pathway or gene names}

\item{\code{goenv}}{environment mapping pathways to genes}

\item{\code{renv}}{global environment}

\item{\code{name}}{app name (needs to be altered only if adding more than one app to the server using the 'server' parameter)}

\item{\code{trim}}{Trim quantity used for Winsorization for visualization}

\item{\code{batch}}{Any batch or other known confounders to be included in the visualization as a column color track}

\item{\code{embedding}}{embedding information}

\item{\code{type}}{Either 'counts' or a name of a 'reduction' in the pagoda2 object}

\item{\code{veloinfo}}{velocity information}
}}


