% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp}
\alias{pagoda2WebApp}
\title{pagoda2WebApp R6 class}
\description{
class to create pagoda2 web applications via a Rook server
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{originalP2object}}{Input pagoda2 object}

\item{\code{name}}{string Display name for the application}

\item{\code{verbose}}{boolean Whether to give verbose output}

\item{\code{cellmetadata}}{metadata associated with pagoda2 object}

\item{\code{mainDendrogram}}{hclust dendrogram of all cells in the pagoda2 object}

\item{\code{geneSets}}{gene sets in the pagoda2 object}

\item{\code{rookRoot}}{Rook server root directory}

\item{\code{appmetadata}}{pagoda2 web application metadata}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pagoda2WebApp$new()}}
\item \href{#method-generateDendrogramOfGroups}{\code{pagoda2WebApp$generateDendrogramOfGroups()}}
\item \href{#method-packCompressInt32Array}{\code{pagoda2WebApp$packCompressInt32Array()}}
\item \href{#method-packCompressFloat64Array}{\code{pagoda2WebApp$packCompressFloat64Array()}}
\item \href{#method-call}{\code{pagoda2WebApp$call()}}
\item \href{#method-readStaticFile}{\code{pagoda2WebApp$readStaticFile()}}
\item \href{#method-updateRookRoot}{\code{pagoda2WebApp$updateRookRoot()}}
\item \href{#method-serializeToStaticFast}{\code{pagoda2WebApp$serializeToStaticFast()}}
\item \href{#method-sparseMatList}{\code{pagoda2WebApp$sparseMatList()}}
\item \href{#method-arrayToJSON}{\code{pagoda2WebApp$arrayToJSON()}}
\item \href{#method-getCompressedEmbedding}{\code{pagoda2WebApp$getCompressedEmbedding()}}
\item \href{#method-serverLog}{\code{pagoda2WebApp$serverLog()}}
\item \href{#method-reducedDendrogramJSON}{\code{pagoda2WebApp$reducedDendrogramJSON()}}
\item \href{#method-cellOrderJSON}{\code{pagoda2WebApp$cellOrderJSON()}}
\item \href{#method-availableAspectsJSON}{\code{pagoda2WebApp$availableAspectsJSON()}}
\item \href{#method-cellmetadataJSON}{\code{pagoda2WebApp$cellmetadataJSON()}}
\item \href{#method-geneInformationJSON}{\code{pagoda2WebApp$geneInformationJSON()}}
\item \href{#method-generateGeneKnnJSON}{\code{pagoda2WebApp$generateGeneKnnJSON()}}
\item \href{#method-generateEmbeddingStructure}{\code{pagoda2WebApp$generateEmbeddingStructure()}}
\item \href{#method-clone}{\code{pagoda2WebApp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$new(
  pagoda2obj,
  appName = "DefaultPagoda2Name",
  dendGroups,
  verbose = 0,
  debug,
  geneSets,
  metadata = metadata,
  innerOrder = NULL,
  orderDend = FALSE,
  appmetadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateDendrogramOfGroups"></a>}}
\if{latex}{\out{\hypertarget{method-generateDendrogramOfGroups}{}}}
\subsection{Method \code{generateDendrogramOfGroups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$generateDendrogramOfGroups(
  r,
  dendrogramCellGroups,
  innerOrder = NULL,
  orderDend = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-packCompressInt32Array"></a>}}
\if{latex}{\out{\hypertarget{method-packCompressInt32Array}{}}}
\subsection{Method \code{packCompressInt32Array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$packCompressInt32Array(v)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-packCompressFloat64Array"></a>}}
\if{latex}{\out{\hypertarget{method-packCompressFloat64Array}{}}}
\subsection{Method \code{packCompressFloat64Array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$packCompressFloat64Array(v)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call"></a>}}
\if{latex}{\out{\hypertarget{method-call}{}}}
\subsection{Method \code{call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$call(env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-readStaticFile"></a>}}
\if{latex}{\out{\hypertarget{method-readStaticFile}{}}}
\subsection{Method \code{readStaticFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$readStaticFile(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateRookRoot"></a>}}
\if{latex}{\out{\hypertarget{method-updateRookRoot}{}}}
\subsection{Method \code{updateRookRoot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$updateRookRoot(newRoot)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serializeToStaticFast"></a>}}
\if{latex}{\out{\hypertarget{method-serializeToStaticFast}{}}}
\subsection{Method \code{serializeToStaticFast()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$serializeToStaticFast(
  binary.filename = NULL,
  verbose = FALSE,
  verbose.timings = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sparseMatList"></a>}}
\if{latex}{\out{\hypertarget{method-sparseMatList}{}}}
\subsection{Method \code{sparseMatList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$sparseMatList(matsparse)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-arrayToJSON"></a>}}
\if{latex}{\out{\hypertarget{method-arrayToJSON}{}}}
\subsection{Method \code{arrayToJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$arrayToJSON(a = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCompressedEmbedding"></a>}}
\if{latex}{\out{\hypertarget{method-getCompressedEmbedding}{}}}
\subsection{Method \code{getCompressedEmbedding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$getCompressedEmbedding(reduc, embed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serverLog"></a>}}
\if{latex}{\out{\hypertarget{method-serverLog}{}}}
\subsection{Method \code{serverLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$serverLog(message)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reducedDendrogramJSON"></a>}}
\if{latex}{\out{\hypertarget{method-reducedDendrogramJSON}{}}}
\subsection{Method \code{reducedDendrogramJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$reducedDendrogramJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cellOrderJSON"></a>}}
\if{latex}{\out{\hypertarget{method-cellOrderJSON}{}}}
\subsection{Method \code{cellOrderJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$cellOrderJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-availableAspectsJSON"></a>}}
\if{latex}{\out{\hypertarget{method-availableAspectsJSON}{}}}
\subsection{Method \code{availableAspectsJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$availableAspectsJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cellmetadataJSON"></a>}}
\if{latex}{\out{\hypertarget{method-cellmetadataJSON}{}}}
\subsection{Method \code{cellmetadataJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$cellmetadataJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-geneInformationJSON"></a>}}
\if{latex}{\out{\hypertarget{method-geneInformationJSON}{}}}
\subsection{Method \code{geneInformationJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$geneInformationJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateGeneKnnJSON"></a>}}
\if{latex}{\out{\hypertarget{method-generateGeneKnnJSON}{}}}
\subsection{Method \code{generateGeneKnnJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$generateGeneKnnJSON(graph)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateEmbeddingStructure"></a>}}
\if{latex}{\out{\hypertarget{method-generateEmbeddingStructure}{}}}
\subsection{Method \code{generateEmbeddingStructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$generateEmbeddingStructure()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pagoda2WebApp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
