% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_arrayToJSON}
\alias{pagoda2WebApp_arrayToJSON}
\title{pagoda2WebApp_arrayToJSON}
\arguments{
\item{arr}{An array (default=NULL)}
}
\value{
Serialised version of the array in JSON, which includes dimension information as separate fields
}
\description{
Serialise an R array to a JSON object
}
