% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_availableAspectsJSON}
\alias{pagoda2WebApp_availableAspectsJSON}
\title{pagoda2WebApp_availableAspectsJSON}
\value{
JSON with parsed cell order from mainDendrogram$cellorder
}
\description{
Parse pathways from originalP2object$misc$pathwayOD$xv into JSON
}
