% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_call}
\alias{pagoda2WebApp_call}
\title{pagoda2WebApp_call}
\arguments{
\item{env}{The environment argument is a true R environment object which the application is free to modify. Please see the Rook documentation for more details.}
}
\description{
Handle httpd server calls
}
