% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_cellOrderJSON}
\alias{pagoda2WebApp_cellOrderJSON}
\title{pagoda2WebApp_cellOrderJSON}
\value{
JSON with parsed cell order from mainDendrogram$cellorder
}
\description{
Parse mainDendrogram$cellorder into JSON
}
