% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_geneInformationJSON}
\alias{pagoda2WebApp_geneInformationJSON}
\title{pagoda2WebApp_geneInformationJSON}
\value{
JSON with parsed information from genename, dispersion, mean gene expression
}
\description{
Parse originalP2object$misc$varinfo[,c("m","qv")] into JSON
}
