% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_generateDendrogramOfGroups}
\alias{pagoda2WebApp_generateDendrogramOfGroups}
\title{Generate a dendrogram of groups}
\arguments{
\item{dendrogramCellGroups}{Cell groups to input into hclust()}
}
\value{
List of hcGroups, cellorder, and cluster.sizes
}
\description{
Generate a dendrogram of groups
}
