% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_generateEmbeddingStructure}
\alias{pagoda2WebApp_generateEmbeddingStructure}
\title{pagoda2WebApp_generateEmbeddingStructure}
\value{
List with embeddings
}
\description{
Generate information about the embeddings we are exporting
}
