% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_generateGeneKnnJSON}
\alias{pagoda2WebApp_generateGeneKnnJSON}
\title{pagoda2WebApp_generateGeneKnnJSON}
\arguments{
\item{graph}{Input graph}
}
\value{
JSON with gene kNN network
}
\description{
Generate a JSON list representation of the gene kNN network
}
