% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_getCompressedEmbedding}
\alias{pagoda2WebApp_getCompressedEmbedding}
\title{pagoda2WebApp_getCompressedEmbedding}
\arguments{
\item{reduc}{reduction}

\item{embed}{embedding}
}
\value{
compressed embedding as JSON
}
\description{
Compress the embedding
}
