% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_packCompressFloat64Array}
\alias{pagoda2WebApp_packCompressFloat64Array}
\title{pagoda2WebApp_packCompressFloat64Array}
\arguments{
\item{v}{float64 array}
}
\value{
compressed array
}
\description{
Compress float64 array
}
