% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_packCompressInt32Array}
\alias{pagoda2WebApp_packCompressInt32Array}
\title{pagoda2WebApp_packCompressInt32Array}
\arguments{
\item{v}{int32 array}
}
\value{
compressed array
}
\description{
Compress int32 array
}
