% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_readStaticFile}
\alias{pagoda2WebApp_readStaticFile}
\title{pagoda2WebApp_readStaticFile}
\arguments{
\item{filename}{path to filename}
}
\value{
Content to display or error page
}
\description{
Read a static file from the filesystem, and put in the response
}
