% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_reducedDendrogramJSON}
\alias{pagoda2WebApp_reducedDendrogramJSON}
\title{pagoda2WebApp_reducedDendrogramJSON}
\value{
JSON with parsed dendrogram
}
\description{
Parse dendrogram into JSON
}
