% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_serializeToStaticFast}
\alias{pagoda2WebApp_serializeToStaticFast}
\title{pagoda2WebApp_serializeToStaticFast}
\arguments{
\item{binary.filename}{path to binary file (default=NULL)}

\item{verbose}{boolean Whether to give verbose output (default=FALSE)}
}
\value{
static file written by WriteListToBinary(expL=exportList, outfile=binary.filename, verbose=verbose)
}
\description{
Convert serialized file to static file
}
