% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_serverLog}
\alias{pagoda2WebApp_serverLog}
\title{pagoda2WebApp_serverLog}
\arguments{
\item{message}{Message to output for the console}
}
\value{
printed message
}
\description{
Logging function for console
}
