% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\name{pagoda2WebApp_sparseMatList}
\alias{pagoda2WebApp_sparseMatList}
\title{pagoda2WebApp_sparseMatList}
\arguments{
\item{matsparse}{Sparse matrix}
}
\value{
List with slots i, p, x
}
\description{
Create simple List from sparse Matrix with Dimnames as JSON
}
