% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.scoring.R
\name{plotAllWithValues}
\alias{plotAllWithValues}
\title{Plot multiple pagoda2 objects with values, optionally saving to file}
\usage{
plotAllWithValues(p2.objs, values, filename = NULL, panel.size = 600,
  mark.cluster.cex = 0.8, verbose = F)
}
\arguments{
\item{p2.objs}{list of pagoda2 objects to plot}

\item{values}{the values to plot}

\item{filename}{name of file to save the plot in, if NULL plot to current device}

\item{panel.size}{the size of the panel for saving to file}

\item{mark.cluster.cex}{size of cluster labels}

\item{verbose}{logical verbosity}
}
\description{
Plot multiple pagoda2 objects with values, optionally saving to file
}
