% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{plotEmbeddingColorByP2Selection}
\alias{plotEmbeddingColorByP2Selection}
\title{Plot a specified pagoda2 embedding according to the selection object}
\usage{
plotEmbeddingColorByP2Selection(
  emb,
  sel,
  show.unlabelled = TRUE,
  show.labels = TRUE,
  label.size = 3,
  point.size = 2,
  show.guides = T,
  alpha = 0.7,
  show.axis = TRUE
)
}
\arguments{
\item{emb}{An embedding from a pagoda2 object}

\item{sel}{a pagoda2 web selection object}

\item{show.unlabelled}{display the plots without any associated selection as grey}

\item{show.labels}{display the labels at the center of each cluster}
}
\value{
a ggplot2 object
}
\description{
Plots the specified pagoda2 embedding according to the specified selection object
}
