% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{plotJointClustering}
\alias{plotJointClustering}
\title{Side by side plot jointly called clusters from getJointClustering()}
\usage{
plotJointClustering(r.n, cl, alpha = 0.3, main = NULL)
}
\arguments{
\item{r.n}{the named list of the pagoda2 jointly called clusters}

\item{cl}{the clusters factor as returned by getJointClustering()}
}
\description{
Side by side plot jointly called clusters from getJointClustering()
}
