% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{plotJointClusters}
\alias{plotJointClusters}
\title{Side by side plot jointly called clusters from get.joint.clusters()}
\usage{
plotJointClusters(r.n, cl)
}
\arguments{
\item{r.n}{the named list of the pagoda2 jointly called clusters}

\item{cl}{the clusters factor as returned by get.joint.clusters()}
}
\description{
Side by side plot jointly called clusters from get.joint.clusters()
}
