% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{plotMulticlassified}
\alias{plotMulticlassified}
\title{Plot multiclassified cells per selection as a percent barplot}
\usage{
plotMulticlassified(sel)
}
\arguments{
\item{sel}{pagoda2 selection object}
}
\value{
ggplot2 object
}
\description{
Plot multiclassified cells per selection as a percent barplot
}
