% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.scoring.R
\name{plotOneWithValues}
\alias{plotOneWithValues}
\title{Plot the embedding of a pagoda2 object with the given values}
\usage{
plotOneWithValues(
  p2obj,
  values,
  title = "",
  type = "PCA",
  embeddingType = "tSNE"
)
}
\arguments{
\item{p2obj}{the pagoda2 object}

\item{values}{the values to plot}

\item{title}{character Title for the plot (default="")}

\item{type}{character Type reduction on which the embedding is based on (default="PCA")}

\item{embeddingType}{characer Type of embedding to plot (default="tSNE")}
}
\value{
NULL, simply updates p2obj$plotEmbedding()
}
\description{
Plot the embedding of a pagoda2 object with the given values
}
