% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{plotSelectionOverlaps}
\alias{plotSelectionOverlaps}
\title{Get a dataframe and plot summarising overlaps between selection of a pagoda2 selection object
ignore self overlaps}
\usage{
plotSelectionOverlaps(sel)
}
\arguments{
\item{sel}{a pagoda2 selection object}
}
\value{
a list that contains a ggplot2 object and a datatable with the overlaps data
}
\description{
Get a dataframe and plot summarising overlaps between selection of a pagoda2 selection object
ignore self overlaps
}
