% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read10Xmatrices}
\alias{read10Xmatrices}
\title{Read multiple 10x matrices into a single sparse array. 
Given a named list of paths of 10X matrices return a single large matrix
with all the data and cell prefixed with the corresponding sample name}
\usage{
read10Xmatrices(
  paths,
  min.common.genes = 1000,
  common.genes = FALSE,
  merge = FALSE,
  prefix.cells = FALSE,
  prefix.sep = "_"
)
}
\arguments{
\item{paths}{named vector of location of the data (readable by read10Xmatrix())}

\item{min.common.genes}{numeric Minimum number of common genes to allow (default=1000)}

\item{common.genes}{boolean Subset all matrices to common genes, required for merge (default=FALSE)}

\item{merge}{boolean Merge all the matrices to one, requires common.genes and prefix.cells (default=FALSE)}

\item{prefix.cells}{boolean Prefix all cells with the name of the respective path in paths (default=FALSE)}

\item{prefix.sep}{character Separator for prefix of cells (default='_')}
}
\value{
Sparse matrix of the Matrix package that contains all the data prefixes by the corresponding sample name
}
\description{
Read multiple 10x matrices into a single sparse array. 
Given a named list of paths of 10X matrices return a single large matrix
with all the data and cell prefixed with the corresponding sample name
}
