% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{readPagoda2SelectionAsFactor}
\alias{readPagoda2SelectionAsFactor}
\title{Read a pagoda2 cell selection file and return it as a factor
while removing any mutliclassified cells}
\usage{
readPagoda2SelectionAsFactor(filepath, use.internal.name = FALSE)
}
\arguments{
\item{filepath}{name of the selection file}

\item{use.internal.name}{boolean Use field 'internal.name' as factor names (default=FALSE). Passed to factorFromP2Selection}
}
\value{
a name factor with the membership of all the cells that are not multiclassified
}
\description{
Read a pagoda2 cell selection file and return it as a factor
while removing any mutliclassified cells
}
