% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{removeSelectionOverlaps}
\alias{removeSelectionOverlaps}
\title{remove cells that are present in more than one selection from all selections}
\usage{
removeSelectionOverlaps(selections)
}
\arguments{
\item{selection}{a pagoda2 selections list}
}
\value{
a new list with the duplicated cells removed
}
\description{
remove cells that are present in more than one selection from all the
selections they are in
}
