% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-scoring.R
\name{score.cells.puram}
\alias{score.cells.puram}
\title{Puram, Bernstein (Cell, 2018)
Score cells as described in Puram, Bernstein (Cell, 2018)}
\usage{
score.cells.puram(data, signature, correct = TRUE, show.plot = FALSE, ...)
}
\arguments{
\item{data}{matrix of expression, rows are cell, columns are genes}

\item{signature}{character vector The signature to evaluate, a character vector of genes}

\item{correct}{boolean Perform background correction by getting a semi-random geneset (default=TRUE)}

\item{show.plot}{boolean If corrected values are calculated show plot of corrected vs original scores (default=FALSE)}

\item{...}{options for get.control.geneset()}
}
\value{
a score for each cell
}
\description{
Puram, Bernstein (Cell, 2018)
Score cells as described in Puram, Bernstein (Cell, 2018)
}
