% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show.app}
\alias{show.app}
\title{Directly open the 'pagoda2' web application and view the 'p2web' application object from our R session}
\usage{
show.app(app, name, port, ip, browse = TRUE, server = NULL)
}
\arguments{
\item{app}{'pagoda2' application object}

\item{name}{character Name of the application to view}

\item{port}{numeric Port number}

\item{ip}{numeric IP address}

\item{browse}{boolean Whether to load the app into an HTML browser (default=TRUE)}

\item{server}{server If NULL, will grab server with get.scde.server(port=port, ip=ip) (derfault=NULL)}
}
\value{
application within browser
}
\description{
Directly open the 'pagoda2' web application and view the 'p2web' application object from our R session
}
