% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sn}
\alias{sn}
\title{Set names equal to values, a stats::setNames wrapper function}
\usage{
sn(x)
}
\arguments{
\item{x}{an object for which names attribute will be meaningful}
}
\value{
An object with names assigned equal to values
}
\description{
Set names equal to values, a stats::setNames wrapper function
}
\keyword{internal}
