% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.scoring.R
\name{subset.signature.to.data}
\alias{subset.signature.to.data}
\title{Subset a gene signature to the genes in the given matrix
with optional warning if genes are missing}
\usage{
\method{subset}{signature.to.data}(data, signature, raise.warning = TRUE)
}
\arguments{
\item{data}{the matrix}

\item{signature}{character vector The signature to subset a character vector of genes}

\item{raise.warning}{boolean Warn if genes are missing (default=TRUE)}
}
\description{
Subset a gene signature to the genes in the given matrix
with optional warning if genes are missing
}
