% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-scoring.R
\name{subsetSignatureToData}
\alias{subsetSignatureToData}
\title{Subset a gene signature to the genes in the given matrix
with optional warning if genes are missing}
\usage{
subsetSignatureToData(data, signature, raise.warning = TRUE)
}
\arguments{
\item{data}{matrix}

\item{signature}{character vector The gene signature from which to subset a character vector of genes}

\item{raise.warning}{boolean Warn if genes are missing (default=TRUE)}
}
\value{
The filtered subset of gene signatures
}
\description{
Subset a gene signature to the genes in the given matrix
with optional warning if genes are missing
}
