% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{validateSelectionsObject}
\alias{validateSelectionsObject}
\title{Validates a pagoda2 selection object}
\usage{
validateSelectionsObject(selections)
}
\arguments{
\item{selections}{the pagoda2 selection object to be validated}
}
\value{
a logical value indicating if the object is valid
}
\description{
Validates a pagoda2 selection object
}
