% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{webP2proc}
\alias{webP2proc}
\title{Generate a pagoda2 web object}
\usage{
webP2proc(
  p2,
  additionalMetadata = NULL,
  title = "Pagoda 2",
  n.cores = 20,
  make.go.sets = TRUE,
  make.de.sets = TRUE,
  go.env = NULL,
  make.gene.graph = TRUE,
  appmetadata = NULL
)
}
\arguments{
\item{p2}{pagoda2 object}

\item{additionalMetadata}{pagoda2 web metadata object (default=NULL)}

\item{title}{character string Title for the web app (default='Pagoda 2')}

\item{n.cores}{numeric Number of cores to use (default=20)}

\item{make.go.sets}{boolean Whether GO sets should be made (default=TRUE)}

\item{make.de.sets}{boolean Whether differential expression sets should be made (default=TRUE)}

\item{go.env}{the GO environment used for the overdispersion analysis (default=NULL)}

\item{make.gene.graph}{logical specifying if the gene graph should be make, if FALSE the find similar genes functionality will be disabled on the web app}
}
\value{
a pagoda2 web application
}
\description{
Generate a pagoda2 web object
}
