% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{writePagoda2SelectionFile}
\alias{writePagoda2SelectionFile}
\title{writes a pagoda2 selection object as a p2 pagoda2 selection files}
\usage{
writePagoda2SelectionFile(sel, filepath)
}
\arguments{
\item{sel}{pagoda2 selection object}

\item{filepath}{name of file to write to}
}
\description{
writes a pagoda2 selection object as a p2 selection file that be be
loaded to the web interfact
}
