% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{armaCor}
\alias{armaCor}
\title{armaCor - matrix column correlations. Allows faster matrix correlations with armadillo. Similar to cor() call, will calculate correlation between matrix columns}
\usage{
armaCor(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix with columns as correlations
}
\description{
armaCor - matrix column correlations. Allows faster matrix correlations with armadillo. Similar to cor() call, will calculate correlation between matrix columns
}
