% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{basicP2proc}
\alias{basicP2proc}
\title{Perform basic pagoda2 processing}
\usage{
basicP2proc(cd, n.cores = 20, batch = NULL)
}
\arguments{
\item{n.cores}{number of cores to use}

\item{batch}{optional batch factor}
}
\value{
a new pagoda2 object
}
\description{
adjust variance, calculate pca reduction
make knn graph, identify clusters with infomap, multilevel and walktrap and make
largeVis and tSNE embeddings
}
