% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{basicP2web}
\alias{basicP2web}
\title{Generate a 'pagoda2' web application from a 'Pagoda2' object}
\usage{
basicP2web(p2, app.title = "Pagoda2", extraWebMetadata = NULL, n.cores = 4)
}
\arguments{
\item{p2}{a 'Pagoda2' object}

\item{app.title}{name of application as displayed in the browser title (default='Pagoda2')}

\item{extraWebMetadata}{additional metadata generated by p2.metadata.from.fractor (default=NULL)}

\item{n.cores}{numeric Number of cores to use for differential expression calculation (default=4)}
}
\value{
a 'pagoda2' web object
}
\description{
Generate a 'pagoda2' web application from a 'Pagoda2' object
}
