% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{calcMulticlassified}
\alias{calcMulticlassified}
\title{Returns a list vector with the number of cells that are
present in more than one selections in the provided p2 selection object}
\usage{
calcMulticlassified(sel)
}
\arguments{
\item{sel}{a pagoda2 selection as genereated by readPagoda2SelectionFile}
}
\value{
list vector with the number of cells that are
    present in more than one selections in the provided p2 selection object
}
\description{
Returns a list vector with the number of cells that are
present in more than one selections in the provided p2 selection object
}
