% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{extendedP2proc}
\alias{extendedP2proc}
\title{Perform extended pagoda2 processing. 
Generate organism specific GO environment and calculate pathway overdispersion.}
\usage{
extendedP2proc(p2, organism = "hs")
}
\arguments{
\item{p2}{the pagoda2 object}

\item{organism}{character Organisms hs (Homo Sapiens), mm (M. Musculus, mouse) or dr (D. Rerio, zebrafish) (default='hs')}
}
\value{
list of a pagoda2 object and go.env
}
\description{
Perform extended pagoda2 processing. 
Generate organism specific GO environment and calculate pathway overdispersion.
}
\examples{
\donttest{
cm <- p2data::sample_BM1
counts <- gene.vs.molecule.cell.filter(cm, min.cell.size=500)
counts <- counts[rowSums(counts)>=10,]
rownames(counts) <- make.unique(rownames(counts))
r <- Pagoda2$new(counts,log.scale=TRUE, n.cores=1)
extendedP2proc(r, organism = 'hs')
}

}
