% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{extendedP2proc}
\alias{extendedP2proc}
\title{Perform extended pagoda2 processing. 
Generate organism specific GO environment and calculate pathway overdispersion.}
\usage{
extendedP2proc(p2, organism = "hs")
}
\arguments{
\item{p2}{the pagoda 2 object}

\item{organism}{character Organisms hs (Homo Sapiens), mm (M. Musculus, mouse) or dr (D. Rerio, zebrafish) (default='hs')}
}
\value{
list of a pagoda2 object and go.env
}
\description{
Perform extended pagoda2 processing. 
Generate organism specific GO environment and calculate pathway overdispersion.
}
\examples{
\donttest{
cm <- readRDS(system.file("extdata", "sample_BM1.rds", package="pagoda2"))
p2 <- basicP2proc(cm)
extendedP2proc(p2)
}
}
