% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{extendedP2proc}
\alias{extendedP2proc}
\title{Perform extended pagoda2 processing}
\usage{
extendedP2proc(p2, n.cores = 20, organism = "hs")
}
\arguments{
\item{p2}{the pagoda 2 object}

\item{n.cores}{number of cores to use}

\item{organism}{organisms hs, mm or dr}
}
\value{
a list of a p2 object and a go.env
}
\description{
generate organism specific GO environment and
calculate pathway overdispersion
}
