% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{extendedP2proc}
\alias{extendedP2proc}
\title{Perform extended 'Pagoda2' processing. 
Generate organism specific GO environment and calculate pathway overdispersion.}
\usage{
extendedP2proc(p2, organism = "hs")
}
\arguments{
\item{p2}{the 'Pagoda2' object}

\item{organism}{character Organisms hs (Homo Sapiens), mm (M. Musculus, mouse) or dr (D. Rerio, zebrafish) (default='hs')}
}
\value{
list of a 'Pagoda2' object and go.env
}
\description{
Perform extended 'Pagoda2' processing. 
Generate organism specific GO environment and calculate pathway overdispersion.
}
