% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{factorToP2selection}
\alias{factorToP2selection}
\title{Converts a names factor to a p2 selection object
if colors are provided it assigns those, otherwise uses a rainbow palette}
\usage{
factorToP2selection(cl, col = NULL)
}
\arguments{
\item{cl}{factor}

\item{col}{names vector of colors (default=NULL)}
}
\value{
a p2 selection object (list)
}
\description{
Converts a names factor to a p2 selection object
if colors are provided it assigns those, otherwise uses a rainbow palette
}
