% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{gene.vs.molecule.cell.filter}
\alias{gene.vs.molecule.cell.filter}
\title{Filter cells based on gene/molecule dependency}
\usage{
gene.vs.molecule.cell.filter(countMatrix, min.cell.size = 500,
  max.cell.size = 50000, p.level = min(0.001, 1/ncol(countMatrix)),
  alpha = 0.1, plot = T, do.par = T)
}
\arguments{
\item{countMatrix}{input count matrix to be filtered}

\item{min.cell.size}{min allowed cell size (default 500)}

\item{max.cell.size}{max allowed cell size (default 5e4)}

\item{p.level}{statistical confidence level for deviation from the main trend, used for cell filtering}

\item{alpha}{shading of the confidence band}

\item{plot}{plot the molecule distribution and the gene/molecule dependency fit}

\item{do.par}{reset graphical parameters prior to plotting}
}
\value{
a filtered matrix
}
\description{
filter cells based on the gene/molecule dependency
}
