% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{get.de.geneset}
\alias{get.de.geneset}
\title{Generate differential expression genesets for the web app given a cell grouping by
calculating DE sets between each cell set and everything else}
\usage{
get.de.geneset(pagObj, groups, prefix = "de_")
}
\arguments{
\item{pagObj}{pagoda object}

\item{groups}{named factor to do the de by}

\item{prefix}{chararcter Prefix to assign to genesets generated (default="de_")}
}
\value{
a 'pagoda2' web object
}
\description{
Generate differential expression genesets for the web app given a cell grouping by
calculating DE sets between each cell set and everything else
}
