% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{getColorsFromP2Selection}
\alias{getColorsFromP2Selection}
\title{Retrieves the colors of each selection from a p2 selection object as a names vector of strings}
\usage{
getColorsFromP2Selection(sel)
}
\arguments{
\item{sel}{pagoda2 selection object}
}
\value{
a named vector of hex colours
}
\description{
Retrieves the colors of each selection from a p2 selection object as a names vector of strings
}
