% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{getJointClustering}
\alias{getJointClustering}
\title{Obtain joint clustering between multiple pagoda2 applications}
\usage{
getJointClustering(r.n, k = 30,
  community.detection.method = walktrap.community,
  community.detection.params = list(), min.group.size = 10, ncomps = 100,
  include.sample.internal.edges = TRUE, mnn.edge.weight = 1,
  internal.edge.weight = 1, extra.info = F, neighbourhood.average = TRUE,
  neighbourhood.k = 20, mutualOnly = TRUE, neighbourhood.cleanup = FALSE,
  custom.neighbourhood.cleanup = NULL,
  neighbourhood.cleanup.params = list(), stop.return.graph = FALSE,
  networkPairFunction = NULL)
}
\arguments{
\item{r.n}{a named list of pagoda2 apps}

\item{k}{number of KNN to look for}

\item{community.detection.method}{one of the igraph package community detection methods (or compatible fn)}

\item{min.group.size}{minimum group size to keep after matching}

\item{ncomps}{number of components to use}

\item{include.sample.internal.edges}{augment MNN network with intra app network}

\item{internal.edge.weight}{weigth of edges from intra-sample matches (sample KNN)}

\item{extra.info}{return extra debugging information (modifies return value structure)}

\item{neighbourhood.average}{logical, average cell neighbourhoods before doing MNN lookup}

\item{neighbourhood.k}{number of nearest neighbours to average over}

\item{mutualOnly}{logical if true (defalt) only interapp links that are mutual are used, otherwise the links don'e need to be mutual (experimental)}

\item{mnn.edge.weigth}{weight of edges from inter-sample matches}
}
\value{
a named (by cell name) factor of groups if extra.info is false, otherwise a list the factor and the extra information
}
\description{
obtain joint clustering between multiple pagoda2 applications
by identifying mutual nearest neighbours pairwise and constructing
a network on which traditional cluster identification algorithms are used to identify clusters
}
