% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{getJointClusterMarkerGenes}
\alias{getJointClusterMarkerGenes}
\title{This function identifies marker genes that support joint clustering
among different pagoda2 application}
\usage{
getJointClusterMarkerGenes(applist, jc)
}
\arguments{
\item{applist}{a list of Pagoda2 object}

\item{js}{a factor with joint clusters as returned by getJointClustering()}
}
\value{
a list of markers for each cluster
}
\description{
This function performs differential expression between clusters
identified with getJointClustering() and looks for genes that are consistently
differentially regulated between different application
}
