% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermatch.R
\name{identifyCellsGSVDMNNmulti}
\alias{identifyCellsGSVDMNNmulti}
\title{Classify cells in a pagoda2 application given a set of annotated reference applications}
\usage{
identifyCellsGSVDMNNmulti(referencesets, annotset, clustersOrig)
}
\arguments{
\item{referencesets}{input datasets (p2 objects in a names list)}

\item{annotset}{p2 dataset to be annotated}

\item{clustersOrigin}{a names factor with cluster assignments for referenceset cells}
}
\value{
a named factor with labells for annotset
}
\description{
This function is similar to identifyCellsGSVDMNN() but it can
take as input a list of annotated pagoda2 objects. It then classifies cells
against all applications and summarised the results.
}
