% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{make.p2.app}
\alias{make.p2.app}
\title{Generate a Rook Server app from a pagoda2 object}
\usage{
make.p2.app(r, dendrogramCellGroups, additionalMetadata = list(), geneSets)
}
\arguments{
\item{r}{pagoda2 object}

\item{additionalMetadata}{a list of metadata other than depth, batch and cluster that are automatically added}

\item{geneSets}{a list of genesets to show}

\item{dendrogramCelllGoups}{a named factor of cell groups, used to generate the main dendrogram, limits zoom in}

\item{debug}{build debug app?}
}
\value{
a Rook web app
}
\description{
Contains some code required to convert from the pagoda2 object to the
web object constructor. Advanced users may wish to use the PagodaWebApp
constructor directly
}
