% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.scoring.R
\name{minMaxScale}
\alias{minMaxScale}
\title{Scale the designated values between the range of 0 and 1}
\usage{
minMaxScale(x)
}
\arguments{
\item{x}{values to scale}
}
\value{
the scaled values
}
\description{
Scale the designated values between the range of 0 and 1
}
\examples{
example_matrix =  matrix(rep(c(1:5), 3), 5)
minMaxScale(example_matrix)

}
