% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebHelpers.R
\name{p2.generate.dr.go.web}
\alias{p2.generate.dr.go.web}
\title{Generates zebrafish (Danio rerio) GO annotation for the web object}
\usage{
p2.generate.dr.go.web(gene.names, n.cores = 1)
}
\arguments{
\item{gene.names}{a character vector of genes to include}

\item{n.cores}{numeric Number of cores to use (default=1)}
}
\description{
Generates zebrafish (Danio rerio) GO annotation for the web object
}
\keyword{internal}
