% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{p2.generate.go}
\alias{p2.generate.go}
\title{Generate a GO environment for the organism specified}
\usage{
p2.generate.go(r, organism = NULL)
}
\arguments{
\item{r}{a pagoda2 object}

\item{organism}{the organism, current hs, mm and dr are supported for the human and mouse}
}
\description{
This is a wrapper function for generating a go envirnoment
}
