% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebHelpers.R
\name{p2.generate.go.web}
\alias{p2.generate.go.web}
\title{Generates GO annotation for the web object for any species}
\usage{
p2.generate.go.web(
  gene.names,
  egALIAS2EG = NULL,
  egGO2ALLEGS = NULL,
  n.cores = 1
)
}
\arguments{
\item{gene.names}{a character vector of genes to include}

\item{egALIAS2EG}{(default=NULL)}

\item{egGO2ALLEGS}{(default=NULL)}

\item{n.cores}{numeric Number of cores to use (default=1)}
}
\description{
Generates GO annotation for the web object for any species
}
\keyword{internal}
