% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{p2.metadata.from.factor}
\alias{p2.metadata.from.factor}
\title{Generate a metadata structure for a p2 web object from a named factor}
\usage{
p2.metadata.from.factor(
  metadata,
  displayname = NULL,
  s = 1,
  v = 1,
  start = 0,
  end = NULL,
  pal = NULL
)
}
\arguments{
\item{metadata}{named factor with metadata for individual cells, names must correspond to cells}

\item{displayname}{name to display for the metadata}

\item{s}{s value for rainbow palette}

\item{v}{v value for rainbow palette}

\item{start}{starting value}

\item{end}{ending value}

\item{pal}{optional vector of colours to use, if provided overrides s,v,start and end parameters}
}
\description{
This function will generate a metadata structure that can be passed to
p2 web object constructor as additional metadata given a named factor
}
\examples{
##additionalMetadata <- list()

# Generate metadata the easy way
##additionalMetadata$altCluster <- p2.metadata.from.factor(myPagoda2Object$clusters$PCA[[1]], displayname = 'Infomap')

# Generate metadata by specifying parameters to be passes to the rainbow function
##additionalMetadata$altCluster <- p2.metadata.from.factor(myPagoda2Object$clusters$PCA[[2]], displayname = 'Multilevel', s = 0.7, v = 0.8,start = 0, end = 0.5)

# Generate metadata by specifying a palette
##a <- myPagoda2Object$clusters$PCA[[3]]
##library(colorRamps)
##p1 <- colorRamps::primary.colors(n = nlevels(a))
##names(p1) <- levels(a) # This is optional
##additionalMetadata$altCluster2 <- p2.metadata.from.factor(myPagoda2Object$clusters$PCA[[3]], displayname = 'Walktrap', pal = p1)
}
