% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{p2.metadata.from.factor}
\alias{p2.metadata.from.factor}
\title{Generate a metadata structure for a p2 web object from a named factor}
\usage{
p2.metadata.from.factor(metadata, displayname = NULL, s = 1, v = 1)
}
\arguments{
\item{metadata}{named factor with metadata for individual cells, names must correspond to cells}

\item{displayname}{name to display for the metadata}

\item{s}{s value for rainbow palette}

\item{v}{v value for rainbow palette}
}
\description{
This function will genereate a metadata structure that can be passed to
p2 web object constructor as additional metadata from a named factor of arbitrary values
}
