% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{p2.metadata.from.factor}
\alias{p2.metadata.from.factor}
\title{Generate a list metadata structure that can be passed to a
'pagoda2' web object constructor as additional metadata given a named factor}
\usage{
p2.metadata.from.factor(
  metadata,
  displayname = NULL,
  s = 1,
  v = 1,
  start = 0,
  end = NULL,
  pal = NULL
)
}
\arguments{
\item{metadata}{named factor with metadata for individual cells, names must correspond to cells}

\item{displayname}{character Name to display for the metadata (default=NULL)}

\item{s}{numeric Value for rainbow palette (default=1)}

\item{v}{numeric Value for rainbow palette (default=1)}

\item{start}{numeric Starting value (default=0)}

\item{end}{numeric Ending value (default=NULL)}

\item{pal}{optional vector of colours to use, if provided overrides s,v,start and end parameters (default=NULL)}
}
\value{
list of data, levels, palette to be passed to 'pagoda2' web object constructor
}
\description{
Generate a list metadata structure that can be passed to a
'pagoda2' web object constructor as additional metadata given a named factor
}
