% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p1_liftover.R
\name{pagoda.reduce.loading.redundancy}
\alias{pagoda.reduce.loading.redundancy}
\title{Collapse aspects driven by the same combinations of genes.
Examines PC loading vectors underlying the identified aspects and clusters aspects based
on a product of loading and score correlation (raised to corr.power). Clusters of aspects
driven by the same genes are determined based on the distance.threshold and collapsed.}
\usage{
pagoda.reduce.loading.redundancy(
  tam,
  pwpca,
  clpca = NULL,
  plot = FALSE,
  cluster.method = "complete",
  distance.threshold = 0.01,
  corr.power = 4,
  abs = TRUE,
  n.cores = 1,
  ...
)
}
\arguments{
\item{tam}{output of pagoda.top.aspects()}

\item{pwpca}{output of pagoda.pathway.wPCA()}

\item{clpca}{output of pagoda.gene.clusters() (optional) (default=NULL)}

\item{plot}{boolean Whether to plot the resulting clustering (default=FALSE)}

\item{cluster.method}{string One of the standard clustering methods to be used (default="complete") Note: fastcluster::hclust is used if available or stats::hclust)}

\item{distance.threshold}{numeric Similarity threshold for grouping interdependent aspects (default=0.01)}

\item{corr.power}{numeric Power to which the product of loading and score correlation is raised (default=4)}

\item{abs}{boolean Whether to use absolute correlation (default=TRUE)}

\item{n.cores}{numeric Number of cores to use during processing (default=1)}

\item{...}{additional arguments are passed to the pagoda.view.aspects() method during plotting}
}
\value{
a list structure analogous to that returned by pagoda.top.aspects(), but with addition of a $cnam element containing a list of aspects summarized by each row of the new (reduced) $xv and $xvw
}
\description{
Collapse aspects driven by the same combinations of genes.
Examines PC loading vectors underlying the identified aspects and clusters aspects based
on a product of loading and score correlation (raised to corr.power). Clusters of aspects
driven by the same genes are determined based on the distance.threshold and collapsed.
}
\examples{
\donttest{
## # Example from scde, <https://www.bioconductor.org/packages/release/bioc/html/scde.html>
## data(pollen)  
## cd <- clean.counts(pollen)
## knn <- knn.error.models(cd, k=ncol(cd)/4, n.cores=10, 
##     min.count.threshold=2, min.nonfailed=5, max.model.plots=10)
## varinfo <- pagoda.varnorm(knn, counts = cd, trim = 3/ncol(cd), 
##     max.adj.var = 5, n.cores = 1, plot = FALSE)
## pwpca <- pagoda.pathway.wPCA(varinfo, go.env, n.components=1, 
##     n.cores=10, n.internal.shuffles=50)
## tam <- pagoda.top.aspects(pwpca, return.table = TRUE, 
##     plot=FALSE, z.score=1.96)  # top aspects based on GO only
## tamr <- pagoda.reduce.loading.redundancy(tam, pwpca)
}

}
