% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagoda2WebApp.R
\docType{class}
\name{pagoda2WebApp-class}
\alias{pagoda2WebApp-class}
\alias{pagoda2WebApp}
\title{pagoda2WebApp class to create pagoda2 web applications via a Rook server}
\description{
pagoda2WebApp class to create pagoda2 web applications via a Rook server
}
\section{Fields}{

\describe{
\item{\code{originalP2object}}{Input pagoda2 object}

\item{\code{name}}{string Display name for the application}

\item{\code{verbose}}{integer Verbosity level (default=0).}

\item{\code{mat}}{Embedding}

\item{\code{cellmetadata}}{metadata associated with pagoda2 object}

\item{\code{mainDendrogram}}{hclust dendrogram of all cells in the pagoda2 object}

\item{\code{geneSets}}{gene sets in the pagoda2 object}

\item{\code{rookRoot}}{Rook server root directory}

\item{\code{appmetadata}}{pagoda2 web application metadata}
}}


