% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p1_liftover.R
\name{pathway.pc.correlation.distance}
\alias{pathway.pc.correlation.distance}
\title{Calculate correlation distance between PC magnitudes given a number of target dimensions}
\usage{
pathway.pc.correlation.distance(pcc, xv, n.cores = 1, target.ndf = NULL)
}
\arguments{
\item{pcc}{weighted PC magnitudes e.g. scde::pagoda.pathway.wPCA() gives the weighted PC magnitudes for each gene provided; 
e.g. scde::pagoda.gene.clusters() gives the weighted PC magnitudes for de novo gene sets identified by clustering on expression}

\item{xv}{a matrix of normalized aspect patterns (rows: significant aspects, columns: cells)}

\item{n.cores}{numeric Number of cores to use (default=1)}

\item{target.ndf}{numeric Target dimensions (default=NULL)}
}
\value{
correlation distance matrix, akin to stats dist
}
\description{
Calculate correlation distance between PC magnitudes given a number of target dimensions
}
