% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.scoring.R
\name{plotOneWithValues}
\alias{plotOneWithValues}
\title{Plot the embedding of a pagoda2 object with the given values}
\usage{
plotOneWithValues(p2obj, values, title = "", type = "PCA",
  embeddingType = "tSNE")
}
\arguments{
\item{p2obj}{the pagoda2 object}

\item{values}{the values to plot}

\item{title}{title for the plot}

\item{type}{the type reduction on which the embedding is based on}

\item{embeddingType}{the type of embedding to plot}
}
\description{
Plot the embedding of a pagoda2 object with the given values
}
