% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{removeSelectionOverlaps}
\alias{removeSelectionOverlaps}
\title{Remove cells that are present in more than one selection from all the
selections they are in}
\usage{
removeSelectionOverlaps(selections)
}
\arguments{
\item{selections}{a pagoda2 selections list}
}
\value{
a new list with the duplicated cells removed
}
\description{
Remove cells that are present in more than one selection from all the
selections they are in
}
