% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.scoring.R
\name{score.cells.nb0}
\alias{score.cells.nb0}
\title{Score cells by getting mean expression of genes in signatures}
\usage{
score.cells.nb0(data, signature)
}
\arguments{
\item{data}{matrix}

\item{signature}{the genes in the signature}
}
\value{
cell scores
}
\description{
Score cells by getting mean expression of genes in signatures
}
