% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.scoring.R
\name{score.cells.nb1}
\alias{score.cells.nb1}
\title{Score cells after standardising the expression of each gene removing outliers}
\usage{
score.cells.nb1(data, signature, quantile.cutoff = 0.01)
}
\arguments{
\item{data}{matrix of expression, rows are cell, columns are genes}

\item{signature}{a character vector of genes to use in the signature}

\item{quantile.cutoff}{numeric The quantile extremes to trim before plotting (default=0.0.1)}
}
\value{
The filtered subset of gene signatures
}
\description{
Score cells after standardising the expression of each gene removing outliers
}
\keyword{internal}
