% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{webP2proc}
\alias{webP2proc}
\title{Generate a pagoda2 web object}
\usage{
webP2proc(p2, additionalMetadata = NULL, title = "Pagoda 2", n.cores = 20,
  make.go.sets = TRUE, make.de.sets = TRUE, organism = "hs",
  make.gene.graph = TRUE)
}
\arguments{
\item{p2}{a pagoda2 object}

\item{additionalMetadata}{a pagoda2 web metadata object}

\item{title}{title for the web app}

\item{n.cores}{number of cores to use}

\item{make.go.sets}{logical specifying if go sets hould be made}

\item{make.de.sets}{logical specifying if differential expression sets should be made}

\item{organism}{organisms hs or mm}

\item{make.gene.graph}{logical specifying if the gene graph should be make, if FALSE the find similar genes functionality will be disabled on the web app}
}
\value{
a pagoda2 web application
}
\description{
Generate a pagoda2 web object
}
