% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagodaWebSelections.R
\name{writeGenesAsPagoda2Selection}
\alias{writeGenesAsPagoda2Selection}
\title{Writes a list of genes as a gene selection that can be loaded in the web interface}
\usage{
writeGenesAsPagoda2Selection(name, genes, filename)
}
\arguments{
\item{name}{the name of the selection}

\item{genes}{a string vector of the gene names}

\item{filename}{the filename to save to}
}
\value{
NULL, writes to filepath the list of genes as a gene selection that can be loaded in the web interface
}
\description{
Writes a list of genes as a gene selection that can be loaded in the web interface
}
